+++ near-sdk-macros/src/core_impl/abi/abi_generator.rs
@@ -23,13 +23,19 @@ pub fn generate(i: &ItemImplInfo) -> TokenStream2 {
         #[cfg(not(target_arch = "wasm32"))]
         const _: () = {
             #[no_mangle]
-            pub fn #near_abi_symbol() -> near_sdk::__private::ChunkedAbiEntry {
+            pub extern "C" fn #near_abi_symbol() -> (*const u8, usize) {
                 let mut gen = near_sdk::__private::schemars::gen::SchemaGenerator::default();
                 let functions = vec![#(#functions),*];
-                near_sdk::__private::ChunkedAbiEntry::new(
-                    functions,
-                    gen.into_root_schema_for::<String>()
-                )
+                let mut data = std::mem::ManuallyDrop::new(
+                    near_sdk::serde_json::to_vec(&near_sdk::__private::ChunkedAbiEntry::new(
+                        functions,
+                        gen.into_root_schema_for::<String>(),
+                    ))
+                    .unwrap(),
+                );
+                data.shrink_to_fit();
+                assert!(data.len() == data.capacity());
+                (data.as_ptr(), data.len())
             }
         };
     }
@@ -81,13 +87,27 @@ impl ImplItemMethodInfo {
             Some(doc) => quote! { Some(#doc.to_string()) },
             None => quote! { None },
         };
-        let is_view = matches!(&self.attr_signature_info.method_type, &MethodType::View);
-        let is_init = matches!(
-            &self.attr_signature_info.method_type,
-            &MethodType::Init | &MethodType::InitIgnoreState
-        );
-        let AttrSigInfo { is_payable, is_private, is_handles_result, .. } =
-            self.attr_signature_info;
+        let mut modifiers = vec![];
+        let kind = match &self.attr_signature_info.method_type {
+            &MethodType::View => quote! { near_sdk::__private::AbiFunctionKind::View },
+            &MethodType::Regular => {
+                quote! { near_sdk::__private::AbiFunctionKind::Call }
+            }
+            &MethodType::Init | &MethodType::InitIgnoreState => {
+                modifiers.push(quote! { near_sdk::__private::AbiFunctionModifier::Init });
+                quote! { near_sdk::__private::AbiFunctionKind::Call }
+            }
+        };
+        if self.attr_signature_info.is_payable {
+            modifiers.push(quote! { near_sdk::__private::AbiFunctionModifier::Payable });
+        }
+        if self.attr_signature_info.is_private {
+            modifiers.push(quote! { near_sdk::__private::AbiFunctionModifier::Private });
+        }
+        let modifiers = quote! {
+            vec![#(#modifiers),*]
+        };
+        let AttrSigInfo { is_handles_result, .. } = self.attr_signature_info;
 
         let mut params = Vec::<TokenStream2>::new();
         let mut callbacks = Vec::<TokenStream2>::new();
@@ -214,10 +234,8 @@ impl ImplItemMethodInfo {
              near_sdk::__private::AbiFunction {
                  name: #function_name_str.to_string(),
                  doc: #function_doc,
-                 is_view: #is_view,
-                 is_init: #is_init,
-                 is_payable: #is_payable,
-                 is_private: #is_private,
+                 kind: #kind,
+                 modifiers: #modifiers,
                  params: #params,
                  callbacks: vec![#(#callbacks),*],
                  callbacks_vec: #callback_vec,