+++ near-sdk/compilation_tests/all.rs
@@ -20,4 +20,6 @@ fn compilation_tests() {
     t.pass("compilation_tests/borsh_storage_key_generics.rs");
     t.pass("compilation_tests/function_error.rs");
     t.pass("compilation_tests/enum_near_bindgen.rs");
+    t.pass("compilation_tests/schema_derive.rs");
+    t.compile_fail("compilation_tests/schema_derive_invalids.rs");
 }