+++ near-sdk/src/lib.rs
@@ -5,12 +5,15 @@
 #[cfg(test)]
 extern crate quickcheck;
 
+#[cfg(all(feature = "unstable", feature = "abi"))]
+pub use near_sdk_macros::NearSchema;
 pub use near_sdk_macros::{
-    ext_contract, near_bindgen, BorshStorageKey, FunctionError, PanicOnDefault,
+    ext_contract, near_bindgen, BorshStorageKey, EventMetadata, FunctionError, PanicOnDefault,
 };
 
 pub mod store;
 
+#[cfg(feature = "legacy")]
 pub mod collections;
 mod environment;
 pub use environment::env;